contact-form-validation - refactors form to separate directory
